Tongues, settings, and meditation
Vampires no longer have their jaws replaced with fangs. Due to an existing RimWorld bug, this caused tongues to be removed (yes really) which caused issues with talking and treated vampires as deformed. Tongues are now replaced instead for a combined "tongue and fangs" hediff.

Added vampire settings for daylight damage and daylight slumber to the mod settings menu.

Vampire meditation should be working again. Thanks to Jon in our Discord for making this possible.